Move hasAcceptableValidity
